[Feature] Fix order of pre and postfilters
Prefilters are executed in order inverse to their priorities, e.g.
prefilter with priority 10 will be called before prefilter witj priority
0.

Postfilters are executed in the opposite order: so postfilter with
priority 10 will be executed after postfilter with priority 0.

It is also possible to specify negative priorities for pre and post
filters to inverse this logic.