[Fix] Fix vertical tab handling in libucl
626c7a170f73eb17efb084be49da3b30fe773a61 is an incomplete fix and emits
invalid JSON.

The vertical tab has a short escape in C but not JSON, so we should
emit the long escape. (libucl won't choke on \v in UCL input but
it doesn't properly round-trip: 'foo\vbar' will be parsed into
'foovbar'.)

libucl has an option to escape strings during parsing, so I modified
that in a similar fashion to 626c7a17.