[Minor] Fuzzy upstream error threshold config
It would be nice to be able to configure `max_errors` and `revive_time`
for fuzzy check module for low traffic systems.

When traffic volume is relatively low an error rate builds at very slow
rate and it would take a while to reach default 40% threshold. At the
same time waiting for 60s sometimes could be unreasonable for dynamic
environments to re-resolve rspamd fuzzy storage.