[Fix] Fix a failure calcuating URL reputation.
`url_reputation_filter` was attempting to lookup the token key returned
in the `get_token` callback in an integer-indexed table of requests.
Change to looking up the request by its index.

Commit 73d2cee changed from using `gen_url_queries` to
`lua_util.extract_specific_urls` in the assignment of `requests`.
The latter just returns a list of URLs. The former returns the TLD along
with the number of hits. The URL score calculation uses the number of
hits. Revert to using `gen_url_queries`.