dkim: Enforce a space character after comma in the recipient list
A Cc list like:

  cc:Person A <a@example.org>,Person B <b@example.org>

will be signed as such. The problem is that there should be a white space
after the comma. If the line gets (later) reformed such as

  Cc: Person A <a@example.org>,
  \TABPerson B <b@example.org>

then line is interpreted as
  cc: Person A <a@example.org>, Person B <b@example.org>

which is different and the verification will failed.

In RFC 6376 section 3.4.2. (The "relaxed" Header Canonicalization
Algorithm) it is described how to unfold/replace white space
characters. It refers to RFC 5322 to how unfold continuation lines. I've
been looking for official behaviour but couldn't find. The space after
the comma is what I observed the missing space being the exception.

Enforce a space after a comma. Ignore this if the comma is found within
a quote block ( started by ' or ") or if the comma is escaped (by \).

Issue: #3842

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>