[Fix] Remove check for a score with no symbol being registered
This check is redundant and has proven to confuse users who want to disable some rules.
Indeed, if we are looking for a way to hint a user that the registered symbol `example`
does not match with the scored symbol `exmaple` (e.g. incorrect spelling), then this
error will be detected when we match symcache entries and the scores. The opposite check
is useless to prevent such an error. And since it is much harder to define a dangling score
with no real symbol, we can just ignore this possibility in general, since the FP for this check,
specifically when some rule got intentionally disabled but a score is not removed just annoys
users for no good reason.